Improve markdown spacing
This is just correcting some minor whitespace issues so parsing is less
ambiguous.

Markdown really wants to have an extra blank line to delineate blocks
from one another, because some parsers will output different results
without. For example, it is possible for some parsers to concatenate
blocks together because the double newline is expected to denote the end
of a `<p>` tag. Without the extra newline, some blocks may be joined
with the above block. This shouldn't occur in most cases because
generally a paragraph is also auto-closed by md parsers when they
encounter new block opening syntax, e.g. triple backticks "```" to open
a code block.

It's not really necessary, as this one was parsing fine on Github, but
my neovim was full of lint warnings which are easy to correct and I need
to work on this file a bit more in future commits.