Correct ordered lists
In markdown, ordered lists are denoted by a prefix of any numeric digit
and a period. It is not necessary to manually count up the numeric
prefix in the markdown source document. In fact, it is encouraged that
you use only `1. ` as the notation for ordered list items.

```
1. OL item 1
1. OL item 2
1. OL item 3
...
```

When transformed into HTML, the parser will automatically increment list
items for you. This allows you to re-order list items without having to
manually manage the index as you go.

I don't personally love this bit of magic, but it's useful in this
circumstance because it helps us to realize that the Installation
section does not contain of a single list containing 2 items. Instead,
this section contains 2 paragraphs which resemble ordered list items.

I've taken the liberty of changing that section to include two `h3`
elements instead.